pyproject.toml
[project]
name = "openai-agents-redis-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "colorama>=0.4.6",
    "openai>=1.97.0",
    "openai-agents>=0.2.2",
    "openai-agents-redis",
    "rich>=14.0.0",
]

[dependency-groups]
dev = [
    "pyclean>=3.1.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.4",
]

[tool.uv.sources]
openai-agents-redis = { path = "lib/openai_agents_redis-0.1.0-py3-none-any.whl" }
